* FIX: Fixed building utils * EDIT: Replaced std::bind2nd by std::bind * EDIT: Replaced std::random_shuffle by std::shuffle

+++ code/engine.vc2008/xrNetServer/NET_Client.cpp
@@ -185,7 +185,7 @@ NET_Packet*		INetQueue::Retreive	()
 	else
 	{
 		u32 tmp_time = GetTickCount()-60000;
-		u32 size = unused.size();
+		size_t size = unused.size();
 		if ((LastTimeCreate < tmp_time) &&  (size > 32))
 		{
 			xr_delete(unused.back());
